+++ core/model/authman.go
@@ -14,6 +14,8 @@
 
 package model
 
+import "strings"
+
 // AuthmanSubject contains user name and user email
 type AuthmanSubject struct {
 	SourceID        string   `json:"sourceId"`
@@ -94,17 +96,7 @@ type АuthmanUserResponse struct {
 // АuthmanGroupsResponse Authman groups response wrapper
 type АuthmanGroupsResponse struct {
 	WsFindGroupsResults struct {
-		GroupResults []struct {
-			Extension        string `json:"extension"`
-			DisplayName      string `json:"displayName"`
-			UUID             string `json:"uuid"`
-			Description      string `json:"description"`
-			Enabled          string `json:"enabled"`
-			DisplayExtension string `json:"displayExtension"`
-			Name             string `json:"name"`
-			TypeOfGroup      string `json:"typeOfGroup"`
-			IDIndex          string `json:"idIndex"`
-		} `json:"groupResults"`
+		GroupResults   []AuthmanGroupEntry `json:"groupResults"`
 		ResultMetadata struct {
 			Success       string `json:"success"`
 			ResultCode    string `json:"resultCode"`
@@ -115,4 +107,47 @@ type АuthmanGroupsResponse struct {
 			Millis        string `json:"millis"`
 		} `json:"responseMetadata"`
 	} `json:"WsFindGroupsResults"`
+}
+
+// AuthmanGroupEntry wrapper for single group entry
+type AuthmanGroupEntry struct {
+	Extension        string `json:"extension"`
+	DisplayName      string `json:"displayName"`
+	UUID             string `json:"uuid"`
+	Description      string `json:"description"`
+	Enabled          string `json:"enabled"`
+	DisplayExtension string `json:"displayExtension"`
+	Name             string `json:"name"`
+	TypeOfGroup      string `json:"typeOfGroup"`
+	IDIndex          string `json:"idIndex"`
+}
+
+// HasDescription checks if the group entry has a description
+func (a *AuthmanGroupEntry) HasDescription() bool {
+	return a.Description != ""
+}
+
+// GetGroupPettyTitleAndAdmins Gets the group pretty name and and group admin UINs
+func (a *AuthmanGroupEntry) GetGroupPettyTitleAndAdmins() (string, []string) {
+	if strings.Contains(a.Description, "|") {
+		var first string
+		var adminUINs []string
+		segments := strings.Split(a.Description, "|")
+		if len(segments) > 1 {
+			for index, segment := range segments {
+				if index == 0 {
+					first = strings.ReplaceAll(segment, "\"", "")
+					first = strings.TrimSpace(first)
+				} else {
+					adminUINs = append(adminUINs, strings.ReplaceAll(segment, " ", ""))
+				}
+			}
+			return first, adminUINs
+		}
+	}
+
+	if a.Description != "" {
+		return a.Description, nil
+	}
+	return a.DisplayExtension, nil
 }